refactor: update <GameFieldCell /> dir name
